requirements.txt
streamlit
google-cloud-storage
google-auth
vertexai
pandas
numpy
scikit-learn
pillow
requests
lxml
beautifulsoup4